part7 except lint and css (will add)

+++ src/components/Blog.js
@@ -1,41 +1,58 @@
+import { useDispatch, useSelector } from "react-redux"
+import { addComment, likeBlog, removeBlog } from "../reducers/blogsReducer"
+import { useParams } from 'react-router-dom'
 import { useState } from "react"
 
-const Blog = ({ blog, handleLike, currentUser, handleRemove }) => {
-  const blogStyle = {
-    paddingTop: 2,
-    paddingLeft: 2,
-    border: 'solid',
-    borderWidth: 1,
-    marginBottom: 5
+
+const Blog = () => {
+  const [comment, setComment] = useState("")
+
+  const id = useParams().id
+  const blog = useSelector(state => state.blogs).find(item => id === item.id)
+
+  const dispatch = useDispatch()
+  const currentUser = useSelector(state => state.user)
+
+  if (!blog || !currentUser) {
+    return null
   }
 
-  const [hidden, setHidden] = useState(true)
   const removeable = currentUser.username === blog.user.username
+  const handleRemove = blog => {
+    if (window.confirm(`remove ${blog.title}?`)) {
+      dispatch(removeBlog(blog.id))
+    }
+  }
+
+  const handleComment = event => {
+    event.preventDefault()
+    dispatch(addComment(id, comment))
+    setComment("")
+  }
 
   return(
-    <div style={blogStyle} className="blog">
+    <div>
+      <h2>{blog.title} {blog.author}</h2>
+      <a href={blog.url}>{blog.url}</a>
       <p>
-        {blog.title} {blog.author} 
-        <button onClick={() => setHidden(!hidden)}>
-          {hidden? "view" : "hide"}
-        </button>
+        {blog.likes} 
+        <button onClick={() => dispatch(likeBlog(blog))}>like</button>
       </p>
+      <p>{blog.user.name}</p>
+      {removeable && <button onClick={() => handleRemove(blog)}>remove</button>}
 
-      {
-        hidden || 
-        <>
-          <p>{blog.url}</p>
-          <p>
-            {blog.likes} 
-            <button onClick={() => handleLike(blog)}>like</button>
-          </p>
-          <p>{blog.user.name}</p>
-          {removeable && <button onClick={() => handleRemove(blog)}>remove</button>}
-        </>
-      }
+      <h3>Comments</h3>
+      <form onSubmit={handleComment}>
+            <input type="text" value={comment} name="comment" 
+              onChange={({ target }) => setComment(target.value)}
+            />
+            <button type="submit">submit</button> 
+        </form>
+      <ul>
+        {blog.comments.map((item, id) => <li key={id}>{item}</li>)}
+      </ul>
     </div>  
   )
 }
-
 
 export default Blog
